requirements.txt
tqdm
python-dotenv
anthropic